requirements.txt
distro==1.9.0
et-xmlfile==1.1.0
JPype1==1.5.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tabula-py==2.9.0
tzdata==2024.1
